Refactor IndividualHarborPage's table to a separate component and memoize
Allows the IndividualHarborPageContainer's contents to change without
rerendering the entire table.
