+++ package.json
@@ -20,6 +20,7 @@
     "apollo-boost": "^0.4.7",
     "axios": "^0.19.0",
     "classnames": "^2.2.6",
+    "fast-deep-equal": "^3.1.1",
     "focus-visible": "^5.1.0",
     "formik": "^2.1.4",
     "graphql": "^14.5.8",
