+++ src/features/harborView/HarborViewPageContainer.tsx
@@ -1,29 +1,20 @@
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useQuery } from '@apollo/react-hooks';
-import { useTranslation } from 'react-i18next';
-import { Cell } from 'react-table';
 
-import Table, { Column } from '../../common/table/Table';
 import { INDIVIDUAL_HARBOR_QUERY } from './queries';
 import { INDIVIDUAL_HARBOR } from './__generated__/INDIVIDUAL_HARBOR';
-import { getIndividualHarborData, getBerths, Berth, getPiers, getMaps } from './utils/utils';
+import { getIndividualHarborData, getBerths, getPiers, getMaps } from './utils/utils';
 import HarborViewPage from './HarborViewPage';
 import HarborCard from '../../common/harborCard/HarborCard';
 import LoadingSpinner from '../../common/spinner/LoadingSpinner';
-import { formatDimension } from '../../common/utils/format';
-import PierSelectHeader from './pierSelectHeader/PierSelectHeader';
 import Modal from '../../common/modal/Modal';
 import BerthEditForm from './forms/berthForm/BerthEditForm';
 import BerthCreateForm from './forms/berthForm/BerthCreateForm';
-import HarborViewTableTools from './harborViewTableTools/HarborViewTableTools';
-import BerthDetails from '../../common/berthDetails/BerthDetails';
-import InternalLink from '../../common/internalLink/InternalLink';
 import PierCreateForm from './forms/pierForm/PierCreateForm';
 import PierEditForm from './forms/pierForm/PierEditForm';
-import Chip from '../../common/chip/Chip';
 import HarborEditForm from './forms/harborForm/HarborEditForm';
-import Pagination from '../../common/pagination/Pagination';
+import HarborViewPageTable from './HarborViewPageTable';
 
 const HarborViewPageContainer: React.SFC = () => {
   const [editingHarbor, setEditingHarbor] = useState<boolean>(false);
@@ -33,7 +24,6 @@ const HarborViewPageContainer: React.SFC = () => {
   const [creatingPier, setCreatingPier] = useState<boolean>(false);
   const { id } = useParams<{ id: string }>();
   const { loading, error, data } = useQuery<INDIVIDUAL_HARBOR>(INDIVIDUAL_HARBOR_QUERY, { variables: { id } });
-  const { t, i18n } = useTranslation();
 
   if (loading) return <LoadingSpinner isLoading={loading} />;
   if (error) return <div>Error</div>;
@@ -42,63 +32,6 @@ const HarborViewPageContainer: React.SFC = () => {
 
   if (!harbor) return <div>No data...</div>;
 
-  const columns: Column<Berth>[] = [
-    {
-      Header: t('harborView.tableHeaders.number') || '',
-      accessor: 'number',
-      filter: 'text',
-    },
-    {
-      Header: t('harborView.tableHeaders.identifier') || '',
-      accessor: 'identifier',
-      filter: 'text',
-    },
-    {
-      Cell: ({ cell }: { cell: Cell<Berth> }) => {
-        const isBerthActive = cell.row.original.isActive;
-        if (!isBerthActive) {
-          return <Chip color="red" label={t('harborView.berthProperties.inactive')} />;
-        }
-        const activeLease = cell.row.original.leases?.find((lease) => lease.isActive);
-        if (!activeLease) {
-          return cell.value;
-        }
-        return <InternalLink to={`/customers/${activeLease.customer.id}}`}>{cell.value}</InternalLink>;
-      },
-      Header: t('harborView.tableHeaders.customer') || '',
-      accessor: ({ leases }) => {
-        const activeLease = leases?.find((lease) => lease.isActive);
-        if (!activeLease) return '';
-        return `${activeLease.customer.firstName} ${activeLease.customer.lastName}`;
-      },
-      id: 'leases',
-      filter: 'text',
-    },
-    {
-      Header: t('harborView.tableHeaders.length') || '',
-      accessor: ({ length }) => formatDimension(length, i18n.language),
-      id: 'length',
-      filter: 'text',
-    },
-    {
-      Header: t('harborView.tableHeaders.width') || '',
-      accessor: ({ width }) => formatDimension(width, i18n.language),
-      id: 'width',
-      filter: 'text',
-    },
-    {
-      Header: t('harborView.tableHeaders.depth') || '',
-      accessor: ({ depth }) => formatDimension(depth, i18n.language),
-      id: 'depth',
-      filter: 'text',
-    },
-    {
-      Header: t('harborView.tableHeaders.mooring') || '',
-      accessor: ({ mooringType }) => t([`common.mooringTypes.${mooringType}`, mooringType]),
-      id: 'mooringType',
-      filter: 'text',
-    },
-  ];
   const piers = getPiers(data);
   const berths = getBerths(data);
   const maps = getMaps(data);
@@ -125,47 +58,13 @@ const HarborViewPageContainer: React.SFC = () => {
         }}
         editHarbor={() => setEditingHarbor(true)}
       />
-      <Table
-        data={berths}
-        columns={columns}
-        canSelectRows
-        renderTableToolsTop={(_, setters) => (
-          <HarborViewTableTools
-            onAddBerth={() => setCreatingBerth(true)}
-            onAddPier={() => setCreatingPier(true)}
-            handleGlobalFilter={setters.setGlobalFilter}
-            canAddBerth={canAddBerth}
-          />
-        )}
-        styleMainHeader={false}
-        renderMainHeader={(props) => (
-          <PierSelectHeader
-            piers={piers}
-            onEdit={(pier) => setPierToEdit(pier.id)}
-            selectedPier={piers.find((pier) =>
-              props.state.filters
-                .filter((filter) => filter.id === 'identifier')
-                .find((filter) => filter.value === pier.identifier)
-            )}
-            onPierSelect={(pier) => {
-              props.setFilter('identifier', pier?.identifier);
-            }}
-          />
-        )}
-        renderSubComponent={(row) => (
-          <BerthDetails
-            leases={row.original.leases ?? []}
-            comment={row.original.comment}
-            onEdit={() => setBerthToEdit(row.original.id)}
-          />
-        )}
-        renderPaginator={({ pageIndex, pageCount, goToPage }) => (
-          <Pagination
-            forcePage={pageIndex}
-            pageCount={pageCount || 1}
-            onPageChange={({ selected }) => goToPage(selected)}
-          />
-        )}
+      <HarborViewPageTable
+        berths={berths}
+        piers={piers}
+        onAddPier={() => setCreatingPier(true)}
+        onAddBerth={() => setCreatingBerth(true)}
+        onEditBerth={(berth) => setBerthToEdit(berth.id)}
+        onEditPier={(pier) => setPierToEdit(pier.id)}
       />
       {berthToEdit && (
         <Modal isOpen={!!berthToEdit} toggleModal={() => setBerthToEdit(null)}>
